Вынос логики общей в Смарт компонент и унаследование от него в TaskEdit
